[Refactor] Change the variable name of memberID to id and type from int to Long

+++ Backend/BangWool/src/main/java/Backend/BangWool/member/repository/MemberRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<MemberEntity, Integer> {
+public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
 
     boolean existsByEmail(String email);
     boolean existsByNickname(String nickname);
